chore: Update package dependencies and types
This commit adds new dependencies including axios and react-beautiful-dnd to the project. It also updates the types in the index.ts file to improve clarity and consistency, changing 'Answer' to 'Reponse' and modifying the 'Question' interface to use a new 'QuestionType' type. Additionally, it renames properties in the 'Quiz' interface for better localization.