Updates role styles and contact card formatting.
Updates role styles for various roles to ensure consistent styling. Enhances contact card formatting by improving the display of job titles and civility. Adds support for displaying formations for formateurs on the contact card. Updates styles for the Forgot Password and Reset Password components to align with brand colors and improve UI consistency. Adds a new category for IA in the profile and formation catalogue components.